Echo sample uses mono input to stereo output for lower latency on Oreo

+++ aaudio/echo/src/main/cpp/audio_effect.h
@@ -22,7 +22,7 @@
 
 class AudioEffect {
 public:
-  void process(float *inputBuffer, int32_t samplesPerFrame, int32_t numFrames);
+  void process(int16_t *inputBuffer, int32_t samplesPerFrame, int32_t numFrames);
 };
